Move internal ValidatingCommand to publications package
