+++ packages/publications/src/Commands/MakePublicationTagCommand.php
@@ -5,14 +5,13 @@ declare(strict_types=1);
 namespace Hyde\Publications\Commands;
 
 use Hyde\Console\Commands\Helpers\InputStreamHandler;
-use Hyde\Console\Concerns\ValidatingCommand;
 use Hyde\Framework\Services\DiscoveryService;
 use Hyde\Hyde;
 use Hyde\Publications\Models\PublicationTags;
 use Hyde\Publications\PublicationService;
-use function implode;
 use LaravelZero\Framework\Commands\Command;
 use RuntimeException;
+use function implode;
 use function sprintf;
 
 /**
