+++ packages/publications/src/Commands/ValidatingCommand.php
@@ -2,15 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Console\Concerns;
+namespace Hyde\Publications\Commands;
+
+use Exception;
+use Illuminate\Support\Facades\Validator;
+use LaravelZero\Framework\Commands\Command;
+use RuntimeException;
 
 use function __;
 use function array_merge;
-use Exception;
-use Illuminate\Support\Facades\Validator;
 use function in_array;
-use LaravelZero\Framework\Commands\Command;
-use RuntimeException;
 use function str_ends_with;
 use function ucfirst;
