+++ packages/publications/tests/ValidatingCommandTest.php
@@ -5,18 +5,19 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Feature;
 
 use Closure;
-use Hyde\Console\Concerns\ValidatingCommand;
+use Hyde\Publications\Commands\ValidatingCommand;
 use Hyde\Testing\TestCase;
 use Illuminate\Console\OutputStyle;
 use Illuminate\Support\Facades\Validator;
 use Mockery;
 use PHPUnit\Framework\ExpectationFailedException;
 use RuntimeException;
-use function str_starts_with;
 use Symfony\Component\Console\Question\ChoiceQuestion;
 
+use function str_starts_with;
+
 /**
- * @covers \Hyde\Console\Concerns\ValidatingCommand
+ * @covers \Hyde\Publications\Commands\ValidatingCommand
  */
 class ValidatingCommandTest extends TestCase
 {
